+++ compair/api/lti_consumers.py
@@ -23,7 +23,7 @@ def non_blank_text(value):
 new_consumer_parser = reqparse.RequestParser()
 new_consumer_parser.add_argument('oauth_consumer_key', type=str, required=True)
 new_consumer_parser.add_argument('oauth_consumer_secret', type=str)
-new_consumer_parser.add_argument('user_id_override', type=non_blank_text)
+new_consumer_parser.add_argument('global_unique_identifier_param', type=non_blank_text)
 new_consumer_parser.add_argument('student_number_param', type=non_blank_text)
 
 existing_consumer_parser = new_consumer_parser.copy()
@@ -80,7 +80,7 @@ class ConsumerAPI(Resource):
 
         consumer.oauth_consumer_key = params.get("oauth_consumer_key")
         consumer.oauth_consumer_secret = params.get("oauth_consumer_secret")
-        consumer.user_id_override = params.get("user_id_override")
+        consumer.global_unique_identifier_param = params.get("global_unique_identifier_param")
         consumer.student_number_param = params.get("student_number_param")
 
         try:
@@ -135,7 +135,7 @@ class ConsumerIdAPI(Resource):
 
         consumer.oauth_consumer_key = params.get("oauth_consumer_key")
         consumer.oauth_consumer_secret = params.get("oauth_consumer_secret")
-        consumer.user_id_override = params.get("user_id_override")
+        consumer.global_unique_identifier_param = params.get("global_unique_identifier_param")
         consumer.student_number_param = params.get("student_number_param")
         consumer.active = params.get("active")