+++ compair/models/lti_models/lti_user.py
@@ -7,7 +7,7 @@ from flask import current_app
 
 from . import *
 
-from compair.core import db
+from compair.core import db, display_name_generator
 
 class LTIUser(DefaultTableMixin, UUIDMixin, WriteTrackingMixin):
     __tablename__ = 'lti_user'
@@ -20,13 +20,14 @@ class LTIUser(DefaultTableMixin, UUIDMixin, WriteTrackingMixin):
     lis_person_name_family = db.Column(db.String(255), nullable=True)
     lis_person_name_full = db.Column(db.String(255), nullable=True)
     lis_person_contact_email_primary = db.Column(db.String(255), nullable=True)
+    global_unique_identifier = db.Column(db.String(255), nullable=True)
     compair_user_id = db.Column(db.Integer, db.ForeignKey("user.id", ondelete="CASCADE"),
         nullable=True)
     system_role = db.Column(EnumType(SystemRole), nullable=False)
     student_number = db.Column(db.String(255), nullable=True)
 
     # relationships
-    # user via User Model
+    # compair_user via User Model
     # lti_consumer via LTIConsumer Model
     lti_memberships = db.relationship("LTIMembership", backref="lti_user", lazy="dynamic")
     lti_user_resource_links = db.relationship("LTIUserResourceLink", backref="lti_user", lazy="dynamic")
@@ -38,6 +39,35 @@ class LTIUser(DefaultTableMixin, UUIDMixin, WriteTrackingMixin):
 
     def is_linked_to_user(self):
         return self.compair_user_id != None
+
+    def generate_or_link_user_account(self):
+        from . import SystemRole, User
+
+        if self.compair_user_id == None and self.global_unique_identifier:
+            self.compair_user = User.query \
+                .filter_by(global_unique_identifier=self.global_unique_identifier) \
+                .one_or_none()
+
+            if not self.compair_user:
+                self.compair_user = User(
+                    username=None,
+                    password=None,
+                    system_role=self.system_role,
+                    firstname=self.lis_person_name_given,
+                    lastname=self.lis_person_name_family,
+                    email=self.lis_person_contact_email_primary,
+                    global_unique_identifier=self.global_unique_identifier
+                )
+                if self.compair_user.system_role == SystemRole.student:
+                    self.compair_user.student_number = self.student_number
+
+                # instructors can have their display names set to their full name by default
+                if self.compair_user.system_role != SystemRole.student and self.compair_user.fullname != None:
+                    self.compair_user.displayname = self.compair_user.fullname
+                else:
+                    self.compair_user.displayname = display_name_generator(self.compair_user.system_role.value)
+
+            db.session.commit()
 
     @classmethod
     def get_by_lti_consumer_id_and_user_id(cls, lti_consumer_id, user_id):
@@ -79,10 +109,18 @@ class LTIUser(DefaultTableMixin, UUIDMixin, WriteTrackingMixin):
         lti_user.handle_fullname_with_missing_first_and_last_name()
         lti_user.lis_person_contact_email_primary = tool_provider.lis_person_contact_email_primary
 
+        if lti_consumer.global_unique_identifier_param and lti_consumer.global_unique_identifier_param in tool_provider.launch_params:
+            lti_user.global_unique_identifier = tool_provider.launch_params[lti_consumer.global_unique_identifier_param]
+        else:
+            lti_user.global_unique_identifier = None
+
         if lti_consumer.student_number_param and lti_consumer.student_number_param in tool_provider.launch_params:
             lti_user.student_number = tool_provider.launch_params[lti_consumer.student_number_param]
         else:
             lti_user.student_number = None
+
+        if not lti_user.is_linked_to_user() and lti_user.global_unique_identifier:
+            lti_user.generate_or_link_user_account()
 
         db.session.commit()