+++ compair/static/test/factories/lti_consumer_factory.js
@@ -5,7 +5,8 @@ var ltiConsumerTemplate = {
     "oauth_consumer_key": null,
     "oauth_consumer_secret": null,
     "active": true,
-    "user_id_override": null,
+    "global_unique_identifier_param": null,
+    "student_number_param": null,
     "created": "Mon, 18 Apr 2016 17:38:23 -0000",
     "modified": "Mon, 18 Apr 2016 17:38:23 -0000"
 }