+++ compair/static/test/features/step_definitions/view_lti_consumer.js
@@ -20,7 +20,8 @@ var viewLTIConsumersStepDefinitionsWrapper = function () {
 
         expect(element(by.css("#consumer_oauth_consumer_key")).getText()).to.eventually.equal("consumer_key_1");
         expect(element(by.css("#consumer_oauth_consumer_secret")).getText()).to.eventually.equal("consumer_secret_1");
-        expect(element(by.css("#consumer_user_id_override")).getText()).to.eventually.equal("consumer_user_id_override");
+        expect(element(by.css("#consumer_global_unique_identifier_param")).getText()).to.eventually.equal("consumer_global_unique_identifier_param");
+        expect(element(by.css("#consumer_student_number_param")).getText()).to.eventually.equal("consumer_consumer_student_number_param");
         expect(element(by.model("consumer.active")).isSelected()).to.eventually.equal(true);
 
         expect(element(by.css("#consumer_created")).isPresent()).to.eventually.equal(true);
@@ -32,7 +33,8 @@ var viewLTIConsumersStepDefinitionsWrapper = function () {
 
         expect(element(by.css("#consumer_oauth_consumer_key")).getText()).to.eventually.equal("consumer_key_3");
         expect(element(by.css("#consumer_oauth_consumer_secret")).getText()).to.eventually.equal("consumer_secret_3");
-        expect(element(by.css("#consumer_user_id_override")).getText()).to.eventually.equal("");
+        expect(element(by.css("#consumer_global_unique_identifier_param")).getText()).to.eventually.equal("");
+        expect(element(by.css("#consumer_student_number_param")).getText()).to.eventually.equal("");
         expect(element(by.model("consumer.active")).isSelected()).to.eventually.equal(false);
 
         expect(element(by.css("#consumer_created")).isPresent()).to.eventually.equal(true);